+++ index.html
@@ -231,7 +231,7 @@
   <template id="messages">
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
-  <script type="module" srs="js/main.js"></script>
-  <script srs="js/function.js"></script>
+  <script type="module" srс="js/main.js"></script>
+  <script srс="js/function.js"></script>
 </body>
 </html>
