+++ js/data.js
@@ -45,3 +45,11 @@ export function createPhotosData() {
 
   return photos;
 }
+
+export function generateRandomComment() {
+  const id = getRandomInt(1, 1000);
+  const avatar = `img/avatar-${getRandomInt(1, 6)}.svg`;
+  const message = commentsList[getRandomInt(0, commentsList.length - 1)];
+  const name = 'Пользователь ' + id;
+  return { id, avatar, message, name };
+} // Генерируем случайный комментарий
