+++ js/main.js
@@ -1,5 +1,5 @@
 //Файл - точка входа
-import { generateRandomComment, getRandomInt } from './util.js';
+//import { generateRandomComment, getRandomInt } from './util.js';
 //Возможно лишнее сейчас, тк импорт для createPhotosData написан в data.js
 import { createPhotosData } from './data.js';
